Makes indicator column generation code in preprocess more robust.
It had previously been getting a list of all genotypes by taking all unique genotypes in cells with only one guide assigned. This was a stupid strategy and if a guide only appears in cells in combination with other guides it would never get a column. Now uses the guide metadata directly.